Remove EchoServer dependency from NetSdrClientApp
Removed the project reference to `EchoTcpServer\EchoServer.csproj`
from the `NetSdrClientApp.csproj` file, eliminating the dependency
on the `EchoTcpServer` project. Deleted the `TestForRefference.cs`
file, which included the `TestForRefference` class that relied on
the `EchoServer` class. Cleaned up the `using EchoServer;` directive
and removed all related references, confirming the removal of the
`EchoServer` dependency.